requirements.txt.txt
torch
torchvision
pandas
matplotlib
scikit-learn
Pillow
torchcam
